# services/API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: Apollo Platform API
  description: |
    ## Apollo Criminal Investigation Platform - Comprehensive API Documentation

    The Apollo Platform provides a complete suite of APIs for criminal investigation,
    intelligence gathering, and threat analysis. This documentation covers all available
    endpoints across Node.js microservices and Python intelligence services.

    ### Service Architecture
    - **Node.js Microservices** (Port 3000-3010): Authentication, Operations, Intelligence, Search, Analytics, Notifications
    - **Python Intelligence Server** (Port 8000): OSINT, Blockchain, Facial Recognition, Voice Analysis, GEOINT

    ### Authentication Methods
    - **JWT Bearer Token**: Primary authentication for all protected endpoints
    - **API Key**: Alternative authentication via `X-API-Key` header
    - **MFA**: Multi-factor authentication for sensitive operations

    ### Rate Limiting
    - Free tier: 100 requests/hour
    - Pro tier: 1,000 requests/hour
    - Enterprise: 10,000 requests/hour

    ### Error Handling
    All errors follow a consistent format with error codes and messages.

  version: 2.0.0
  contact:
    name: Apollo Platform Team
    email: dev@apollo-platform.com
    url: https://apollo-platform.com/support
  license:
    name: Proprietary
    url: https://apollo-platform.com/license
  x-logo:
    url: https://apollo-platform.com/logo.png
    altText: Apollo Platform
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    UsernameSearchResult:
      type: object
      properties:
        success:
          type: boolean
        username:
          type: string
        total_platforms_checked:
          type: integer
        platforms_found:
          type: integer
        search_duration_seconds:
          type: number
        results:
          type: array
          items:
            type: object
            properties:
              platform:
                type: string
              url:
                type: string
              status:
                type: string
                enum: [found, not_found, error]
              confidence_score:
                type: number
              response_time_ms:
                type: integer

    # ============ Blockchain Schemas ============
...
paths:
...
  /api/v1/osint/username/search:
    post:
      summary: Username search
      description: Search for username across 400+ platforms
      tags: [OSINT]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  example: "cryptoqueen"
                platforms:
                  type: array
                  items:
                    type: string
                  description: Specific platforms to search (optional)
                timeout:
                  type: integer
                  default: 15
                  description: Timeout per platform in seconds
                check_availability:
                  type: boolean
                  default: false
            example:
              username: "ruja_ignatova"
              timeout: 30
      responses:
        '200':
          description: Username search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsernameSearchResult'
  /api/v1/osint/username/batch-search:
    post:
      summary: Batch username search
      description: Search multiple usernames in parallel
      tags: [OSINT]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [usernames]
              properties:
                usernames:
                  type: array
                  items:
                    type: string
                platforms:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Batch search results
  /api/v1/osint/email/intelligence:
    post:
      summary: Email intelligence
      description: Comprehensive email intelligence gathering
      tags: [OSINT]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Email intelligence results
  /api/v1/osint/phone/lookup:
    post:
      summary: Phone lookup
      description: Phone number intelligence and carrier lookup
      tags: [OSINT]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone_number]
              properties:
                phone_number:
                  type: string
                  example: "+1-555-123-4567"
      responses:
        '200':
          description: Phone lookup results
  /api/v1/osint/domain/scan:
    post:
      summary: Domain scan
      description: Comprehensive domain reconnaissance
      tags: [OSINT]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain]
              properties:
                domain:
                  type: string
                  example: "example.com"
                scan_type:
                  type: string
                  enum: [subdomain_enum, port_scan, vuln_scan, full]
                  default: subdomain_enum
                modules:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Domain scan results
  /api/v1/osint/image/reverse-search:
    post:
      summary: Reverse image search
      description: Search for image across multiple search engines
      tags: [OSINT]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image_url:
                  type: string
                  format: uri
      responses:
        '200':
          description: Reverse image search results
  /api/v1/osint/public-records/search:
    post:
      summary: Public records search
      description: Search public records databases
      tags: [OSINT]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                location:
                  type: string
                record_types:
                  type: array
                  items:
                    type: string
                    enum: [court, property, business, criminal, marriage]
      responses:
        '200':
          description: Public records results
...
security:
  - bearerAuth: []
  - apiKeyAuth: []